bau: Split out ChargeServiceTest
ChargeServiceTest has become too big, so splitting it out.
